Refactor type resolution in ResolvedPath.
