+++ libs/pavex/src/language/resolved_path.rs
@@ -7,7 +7,7 @@ use bimap::BiHashMap;
 use guppy::PackageId;
 use itertools::Itertools;
 use quote::format_ident;
-use rustdoc_types::{Item, ItemEnum};
+use rustdoc_types::Item;
 
 use pavex_builder::RawCallableIdentifiers;
 
@@ -168,60 +168,19 @@ impl ResolvedPath {
         Err(UnknownPath(self.to_owned()))
     }
 
-    /// Find information about the type that the path points at.
+    /// Find information about the type that this path points at.
     pub fn find_type(&self, krate_collection: &mut CrateCollection) -> Result<Item, UnknownPath> {
-        // TODO: remove unwrap here
-        krate_collection
-            .get_or_compute_crate_by_package_id(&self.package_id)
-            .unwrap();
-        let krate = krate_collection.get_crate_by_package_id(&self.package_id);
-        let path_segments: Vec<_> = self
+        let path: Vec<_> = self
             .segments
             .iter()
             .map(|path_segment| path_segment.ident.to_string())
             .collect();
-        if let Ok(type_id) = krate.get_type_id_by_path(&path_segments) {
-            return Ok(krate_collection
-                .get_type_by_global_type_id(type_id)
-                .to_owned());
-        }
-        // The path might be pointing to a method, which is not a type.
-        // We drop the last segment to see if we can get a hit on the struct/enum type
-        // to which the method belongs.
-        if path_segments.len() < 3 {
-            // It has to be at least three segments - crate name, type name, method name.
-            // If it's shorter than three, it's just an unknown path.
-            return Err(UnknownPath(self.to_owned()));
-        }
-        let (method_name, type_path_segments) = path_segments.split_last().unwrap();
-
-        if let Ok(type_id) = krate.get_type_id_by_path(&type_path_segments) {
-            let t = krate_collection.get_type_by_global_type_id(type_id);
-            let impl_block_ids = match &t.inner {
-                ItemEnum::Struct(s) => &s.impls,
-                ItemEnum::Enum(enum_) => &enum_.impls,
-                _ => return Err(UnknownPath(self.to_owned())),
-            };
-            for impl_block_id in impl_block_ids {
-                let impl_block = krate.get_type_by_local_type_id(impl_block_id);
-                if let ItemEnum::Impl(impl_block) = &impl_block.inner {
-                    // We are completely ignoring the bounds attached to the implementation block.
-                    // This can lead to issues: the same method can be defined multiple
-                    // times in different implementation blocks with non-overlapping constraints.
-                    for impl_item_id in &impl_block.items {
-                        let impl_item = krate.get_type_by_local_type_id(impl_item_id);
-                        if impl_item.name.as_ref() == Some(method_name) {
-                            if let ItemEnum::Method(_) = &impl_item.inner {
-                                return Ok(impl_item.to_owned());
-                            }
-                        }
-                    }
-                } else {
-                    unreachable!()
-                }
-            }
-        }
-        Err(UnknownPath(self.to_owned()))
+        krate_collection
+            .get_type_by_local_path(&path, &self.package_id)
+            // TODO: Remove this unwrap
+            .unwrap()
+            .map(ToOwned::to_owned)
+            .map_err(|_| UnknownPath(self.to_owned()))
     }
 
     pub fn render_path(&self, id2name: &BiHashMap<&PackageId, String>) -> String {
